[Behat] Extracted email related steps to EmailContext
